use double for font size instead of float

+++ src/robot_dart/gui_data.hpp
@@ -25,7 +25,7 @@ namespace robot_dart {
             std::vector<std::shared_ptr<simu::TextData>> text_drawings;
 
         public:
-            std::shared_ptr<simu::TextData> add_text(const std::string& text, const Eigen::Affine2d& tf = Eigen::Affine2d::Identity(), Eigen::Vector4d color = Eigen::Vector4d(1, 1, 1, 1), std::uint8_t alignment = (1 | 3 << 3), bool draw_bg = false, Eigen::Vector4d bg_color = Eigen::Vector4d(0, 0, 0, 0.75), float font_size = 28)
+            std::shared_ptr<simu::TextData> add_text(const std::string& text, const Eigen::Affine2d& tf = Eigen::Affine2d::Identity(), Eigen::Vector4d color = Eigen::Vector4d(1, 1, 1, 1), std::uint8_t alignment = (1 | 3 << 3), bool draw_bg = false, Eigen::Vector4d bg_color = Eigen::Vector4d(0, 0, 0, 0.75), double font_size = 28)
             {
                 text_drawings.emplace_back(new TextData{text, tf, color, alignment, draw_bg, bg_color, font_size});
